modified model folder structure
