+++ src/main/java/uk/gov/laa/ccw/controllers/MatterCodesController.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.laa.ccw.mapping.api.MatterCodesResponseMapping;
-import uk.gov.laa.ccw.models.MatterCodes200Response;
+import uk.gov.laa.ccw.models.api.MatterCodes200Response;
 import uk.gov.laa.ccw.services.MatterCodesService;
 
 @Slf4j
