+++ src/main/java/uk/gov/laa/ccw/mapping/api/CaseStagesResponseMapping.java
@@ -1,7 +1,7 @@
 package uk.gov.laa.ccw.mapping.api;
 
 import uk.gov.laa.ccw.models.CaseStage;
-import uk.gov.laa.ccw.models.CaseStages200ResponseCaseStage;
+import uk.gov.laa.ccw.models.api.CaseStages200ResponseCaseStage;
 
 public class CaseStagesResponseMapping {
     public static CaseStages200ResponseCaseStage map(CaseStage inputData) {
