+++ src/main/java/uk/gov/laa/ccw/models/api/CaseStages200ResponseCaseStage.java
@@ -1,16 +1,15 @@
-package uk.gov.laa.ccw.models;
+package uk.gov.laa.ccw.models.api;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class MatterCodes200Response {
-    public List<MatterCodes200ResponseMatterCode> matterCodes;
+public class CaseStages200ResponseCaseStage {
+    public String caseStage;
+    private String description;
 }
