+++ src/main/java/uk/gov/laa/ccw/models/api/HttpError500Response.java
@@ -1,4 +1,4 @@
-package uk.gov.laa.ccw.models;
+package uk.gov.laa.ccw.models.api;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -9,7 +9,6 @@ import lombok.NoArgsConstructor;
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class CaseStages200ResponseCaseStage {
-    public String caseStage;
-    private String description;
+public class HttpError500Response {
+    public String error;
 }
