+++ src/main/java/uk/gov/laa/ccw/models/api/MatterCodes200Response.java
@@ -1,14 +1,16 @@
-package uk.gov.laa.ccw.models;
+package uk.gov.laa.ccw.models.api;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
 @Builder
 @AllArgsConstructor
 @NoArgsConstructor
-public class HttpError404Response {
-    public String error;
+public class MatterCodes200Response {
+    public List<MatterCodes200ResponseMatterCode> matterCodes;
 }
